requirements.txt
Flask
Flask-SocketIO
Flask-Cors
qrcode
requests
kivy
opencv-python
pyzbar
python-socketio
